+++ vck-openid/src/commonTest/kotlin/at/asitplus/wallet/lib/oidvci/OidvciAttestationTest.kt
@@ -89,7 +89,6 @@ class OidvciAttestationTest : FunSpec({
                 token.toHttpHeaderValue(),
                 it,
                 credentialDataProvider = DummyOAuth2IssuerCredentialDataProvider,
-                issueCredential = { IssuerAgent().issueCredential(it) }
             ).getOrThrow()
             val serializedCredential = credential.credentials.shouldNotBeEmpty()
                 .first().credentialString.shouldNotBeNull()
@@ -128,7 +127,6 @@ class OidvciAttestationTest : FunSpec({
                     token.toHttpHeaderValue(),
                     it,
                     credentialDataProvider = DummyOAuth2IssuerCredentialDataProvider,
-                    issueCredential = { IssuerAgent().issueCredential(it) }
                 ).getOrThrow()
             }
         }