+++ vck-openid/src/commonTest/kotlin/at/asitplus/wallet/lib/oidvci/OidvciPreAuthTest.kt
@@ -82,7 +82,6 @@ class OidvciPreAuthTest : FreeSpec({
             authorizationHeader = token.toHttpHeaderValue(),
             params = credentialRequest.first(),
             credentialDataProvider = DummyOAuth2IssuerCredentialDataProvider,
-            issueCredential = { IssuerAgent().issueCredential(it) }
         ).getOrThrow()
         credential.credentials.shouldNotBeEmpty().first().credentialString.shouldNotBeNull()
     }
@@ -115,7 +114,6 @@ class OidvciPreAuthTest : FreeSpec({
                 token.toHttpHeaderValue(),
                 credentialRequest.first(),
                 credentialDataProvider = DummyOAuth2IssuerCredentialDataProvider,
-                issueCredential = { IssuerAgent().issueCredential(it) }
             ).getOrThrow()
                 .credentials.shouldNotBeEmpty().first()
                 .credentialString.shouldNotBeNull()
@@ -154,7 +152,6 @@ class OidvciPreAuthTest : FreeSpec({
             token.toHttpHeaderValue(),
             credentialRequest.first(),
             credentialDataProvider = DummyOAuth2IssuerCredentialDataProvider,
-            issueCredential = { IssuerAgent().issueCredential(it) }
         ).getOrThrow()
             .credentials.shouldNotBeEmpty().first()
             .credentialString.shouldNotBeNull()
@@ -191,7 +188,6 @@ class OidvciPreAuthTest : FreeSpec({
             token.toHttpHeaderValue(),
             credentialRequest,
             credentialDataProvider = DummyOAuth2IssuerCredentialDataProvider,
-            issueCredential = { IssuerAgent().issueCredential(it) }
         ).getOrThrow()
             .credentials.shouldNotBeEmpty()
             .shouldHaveSize(2)